Start new feature branch: operator-buttons
Update HTML file to remove decimal from being grouped with operators. Also update the CSS file to style the decimal properly again.

Select all the operator buttons and create a 'for' loop to add an event listener to each one. Console log the value of the button when a click happens to confirm that the operator variable will be properly set.