Add event listener for number buttons.
Add an event listener for number buttons and pass some of the functions that have already been written for the click event listeners. This feature will enable users to use their keyboard to type in numbers thus allowing for a new way to interact with the application without the use of their mouse.